Make public API use fixed-size integer types as a preparation for network stuff

+++ src/app/video.c
@@ -95,15 +95,15 @@ _video_write_framebuffer(Video *video)
 
     const Settings *const settings = video->settings;
     const GraphicsData *const gfxdata = video->gfxdata;
-    const int width = settings->width;
-    const int height = settings->height;
+    const uint16_t width = settings->width;
+    const uint16_t height = settings->height;
 
     const float *const pxdata = GraphicsData_get_pixel_data(gfxdata);
     uint32_t *const buf = video->image->pixels;
     int i;
 #pragma omp parallel for collapse(2)
     for (i = 0; i < width; ++i) {
-        for (int j = 0; j < height; ++j) {
+        for (uint16_t j = 0; j < height; ++j) {
             buf[j * width + i]
               = video->palette(pxdata[i * height + j], video->palette_params);
         }
